pyproject.toml
[project]
name = "vassar"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "dtaylor-us", email = "derektaylor.us@gmail.com" }]
dependencies = [
    "fastapi>=0.112.1",
    "uvicorn[standard]>=0.30.6",
    "neo4j>=5.23.1",
    "python-fasthtml>=0.4.3",
    "rich>=13.7.1",
    "typer>=0.12.4",
    "instruct-easy @ git+https://github.com/chiefkemist/instruct-easy",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["ruff>=0.6.1"]

[tool.rye.scripts]
web_app_dev = "rye run uvicorn vassar.main:app --port 8001 --reload-dir ."
web_app = "rye run uvicorn vassar.main:app --port 8000"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/vassar"]
